Try split avx2 and naitve code to fix non-avx2 build errors
