+++ src/asm/x86/lrf.rs
@@ -174,6 +174,104 @@ pub unsafe fn get_integral_square_avx2(
   )
 }
 
+pub fn sgrproj_compute_3x3_and_5x5_box_sums(
+  integral_image_buffer: &mut IntegralImageBuffer,
+  integral_image_stride: usize, stripe_w: usize, stripe_h: usize,
+  cpu: CpuFeatureLevel,
+) {
+  if cpu >= CpuFeatureLevel::AVX2 {
+    for r in 1..=2 {
+      let d: usize = r * 2 + 1;
+      // and save the results sum and sum of square in different storage
+      let integral_image_offset =
+        if d == 3 { integral_image_stride + 1 } else { 0 };
+      let iimg =
+        &integral_image_buffer.integral_image[integral_image_offset..];
+      let iimg_sq =
+        &integral_image_buffer.sq_integral_image[integral_image_offset..];
+
+      for y in 0..stripe_h + 2 {
+        if d == 5 && (y & 1) == 1 {
+          continue;
+        }
+        for x in (0..stripe_w + 2).step_by(8) {
+          if x + 8 <= stripe_w + 2 {
+            unsafe {
+              let sum =
+                get_integral_square_avx2(iimg, integral_image_stride, x, y, d);
+              let ssq = get_integral_square_avx2(
+                iimg_sq,
+                integral_image_stride,
+                x,
+                y,
+                d,
+              );
+              // save 4 x u32 sum and sum of square in storage
+              let ptr = if d == 3 {
+                &mut integral_image_buffer.sum_3x3
+              } else {
+                &mut integral_image_buffer.sum_5x5
+              };
+              let ptr_sq = if d == 3 {
+                &mut integral_image_buffer.sum_sq_3x3
+              } else {
+                &mut integral_image_buffer.sum_sq_5x5
+              };
+
+              _mm256_storeu_si256(
+                ptr.as_mut_ptr().add(y * integral_image_stride + x) as *mut _,
+                sum,
+              );
+              _mm256_storeu_si256(
+                ptr_sq.as_mut_ptr().add(y * integral_image_stride + x)
+                  as *mut _,
+                ssq,
+              );
+            }
+          } else {
+            for x2 in x..stripe_w + 2 {
+              let sum = native::get_integral_square(
+                iimg,
+                integral_image_stride,
+                x2,
+                y,
+                d,
+              );
+              let ssq = native::get_integral_square(
+                iimg_sq,
+                integral_image_stride,
+                x2,
+                y,
+                d,
+              );
+              // save sum and sum of square in storage
+              if d == 3 {
+                integral_image_buffer.sum_3x3
+                  [y * integral_image_stride + x2] = sum;
+                integral_image_buffer.sum_sq_3x3
+                  [y * integral_image_stride + x2] = ssq;
+              } else {
+                integral_image_buffer.sum_5x5
+                  [y * integral_image_stride + x2] = sum;
+                integral_image_buffer.sum_sq_5x5
+                  [y * integral_image_stride + x2] = ssq;
+              }
+            }
+          }
+        }
+      }
+    }
+  } else {
+    native::sgrproj_compute_3x3_and_5x5_box_sums(
+      integral_image_buffer,
+      integral_image_stride,
+      stripe_w,
+      stripe_h,
+      cpu,
+    );
+  }
+}
+
 #[allow(unused)]
 #[inline]
 #[target_feature(enable = "avx2")]
